requirements.txt
Flask==3.1.0
PyYAML~=6.0.2
SQLAlchemy~=2.0.38
flask-cors~=6.0.1
coloredlogs~=15.0.1
Flask-JWT-Extended~=4.7.1
pillow~=11.3.0
Werkzeug~=3.1.3
psutil~=7.1.0
Flask-Limiter~=4.0.0
argon2-cffi~=25.1.0
psycopg2~=2.9.11
PyMySQL~=1.1.1
cryptography~=46.0.3